fix: Fix issue that custom games did not load
Fix #38 in a previous refactor the lobby was changed, this
resulted in an issue where the lobby did not correctly communicate
that the deck was changed therefor the game tried to load a null deck.

Increased app version to 0.3.1 to reflect hotfix.

Fix version display on main menu to look better