feat: Rework lobby to be more user friendly
Remove age from player cards, this does not work with a digital game
like this.
Add order number to player cards, this will be used to order them
correctly.
Add buttons to move players up or down in the lobby player list.
Change add player dialog to be embeeded into the player list,
this does give a more smooth experience to add new players.
Change add ai player dialog to be embeeded into the player list,
this does give a more smooth experience to add new ai players.
Add icons to player cards in lobby, this helps to quicky identify
if a player is ai or human.
Add icon to remove player button on lobby screen, this matches
better with the current style.
Fix some issues with focus lost if players are getting deleted,
this should make it more playable with controllers.
Add question mark to player list, this will show an tooltip
how the player order is getting defined.

Add icons to player cards ingame, this helps to quickly identify
if a player is ai or human.