feat: Improve game end screen visuals
Improve the visuals apperance of the game end screen.
Add a new resource containing a color only, this does make sure
that the same highlight color is used for all the highlighted labels.
Add a new shader to change the color of a sprite, this is required to
recolor the game icons.
Add color to player or ai icon if it is the player turn.
Add a new trophy icon to the game end screen to reward players with.

Fix $135