feat: refactor code base
This commit will refactor the memory game scene to extract the deck
building process to it's own node. It will also rework the tutorial to a
more dynamic approach. As a result of that the save file tracking the
tutorial state was also changed.

This should make changes to the whole process much more easy in future
iterations.