feat: rework game art
This commit will add a color theme to the whole game, this will replace
most assets. It will also rework the colors for all UI elements.

On top of that some effects are getting changed to match that new theme.
Therefor some code was changed to use a more dynamic approach.

Add license for new assets to give credit to the creator.
Update readme file to reflect license changes.